Add dropdown for selecting assessor neighborhood to data select control.
Assessor neighborhood (tracked in state as 'featureKey') can now be updated either by clicking on areas on the map or by using the new dropdown, either of which will result in the same behavior. Restyling of the geojson in response to feature selection/deselection has been refactored into stateCallbacks.featureKey to standardize behavior.